Refactor the `JoinPattern` to the `join_pattern` module
For easier navigation, move the `JoinPattern` to a separate module.
Personally I believe having a `types` module is a code smell.
